+++ lib/extension.dart
@@ -1,0 +1,40 @@
+
+extension CustomDateTime on DateTime {
+  int daysInMonth() {
+    if (month > 12) {
+      return 1;
+    }
+    final List<int> monthLength = List.filled(12, 0);
+    monthLength[0] = 31;
+    monthLength[2] = 31;
+    monthLength[4] = 31;
+    monthLength[6] = 31;
+    monthLength[7] = 31;
+    monthLength[9] = 31;
+    monthLength[11] = 31;
+    monthLength[3] = 30;
+    monthLength[8] = 30;
+    monthLength[5] = 30;
+    monthLength[10] = 30;
+
+    if (leapYear(year) == true) {
+      monthLength[1] = 29;
+    } else {
+      monthLength[1] = 28;
+    }
+
+    return monthLength[month - 1];
+  }
+
+  bool leapYear(int year) {
+    bool leapYear = false;
+    final bool leap = (year % 100 == 0) && (year % 400 != 0);
+    if (leap == true) {
+      leapYear = false;
+    } else if (year % 4 == 0) {
+      leapYear = true;
+    }
+
+    return leapYear;
+  }
+}
