+++ lib/widget/calendar/calendar.dart
@@ -1,3 +1,4 @@
+import 'package:fhc_calendar/extension.dart';
 import 'package:fhc_calendar/widget/calendar/calendar_item.dart';
 import 'package:flutter/material.dart';
 
@@ -29,7 +30,8 @@ class Calendar extends StatelessWidget {
   final BoxDecoration? boxDecorationOfItemDay;
   int? dayUpdate;
   int dayForFirstRow = 0;
-  DateTime myDateAfterPlus = DateTime.now().subtract(const Duration(days: 30));
+  DateTime myDateAfterPlus =
+      DateTime.now().subtract(Duration(days: DateTime.now().daysInMonth()));
   final maxRowsNumber = 5;
 
   @override
@@ -58,12 +60,12 @@ class Calendar extends StatelessWidget {
       if (dayUpdate != null) {
         date = DateTime(dateTime.year, dateTime.month, dayUpdate!);
       }
-      if (myDateAfterPlus.month <= date.month && myDateAfterPlus.day != 1) {
-      rows.add(_buildRow(
-        dateTimeDefault: date,
-        context: context,
-      ));
-      }
+      // if (myDateAfterPlus.month <= date.month && myDateAfterPlus.day != 1) {
+        rows.add(_buildRow(
+          dateTimeDefault: date,
+          context: context,
+        ));
+      // }
     }
 
     return rows;
