+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: fhc_calendar
 description: Customize calendar as beautiful as you expected.
-version: 0.0.7
+version: 0.0.8
 homepage: https://github.com/BuiXuanQuan/FHC_Calendar
 
 environment:
